+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Added a new api `linz.api.middleware.getFilters(req)` to get the current list view filters from a post request.
+- Added a new api `linz.api.formtools.list.getFilters(req)` to get the current list view filters from a post request.
 
 ## 1.0.0-18.4.3 (3 June 2019)
