+++ lib/api/middleware/index.js
@@ -3,7 +3,6 @@
 const authenticate = require('./authenticate');
 const error = require('./error');
 const exclusions = require('./exclusions');
-const { getFilters } = require('./filters');
 const logError = require('./log-error');
 const login = require('./login');
 const logout = require('./logout');
@@ -21,7 +20,6 @@ module.exports = {
     authenticate,
     error,
     exclusions,
-    getFilters,
     getModelFromRequest,
     logError,
     login,
